# openapi.yaml
openapi: 3.1.0
info:
  title: Swagger My Digi Network  OpenAPI 3.0
  description: |-
    This is a sample Student Social Network Server based on the OpenAPI 3.0 specification. In this version of the student social network API, we've switched to a design-first approach, ensuring that the API is well-documented and easy to improve collaboratively.

    You can now contribute to this API by suggesting changes to the API definition itself or by modifying the underlying codebase. This collaborative approach allows us to enhance the API continuously, making it easier for students to connect, collaborate, and share resources in a secure and efficient environment.

    Some useful links:
    - The Student Social Network API repository — ([link to the GitHub repository](https://github.com/BumblePlumz/SocialNetwork.git))

    1. User Registration and Authentication: Students can sign up, log in, and manage their profiles.
    2. Articles : Students can post article, comment article and subscribe to other students post.
  termsOfService: http://swagger.io/terms/
  contact:
    email: fake@fake.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /user/photo:
    put:
      summary: Upload or Update Profile Photo
      description: Upload a profile photo either via file upload or base64 encoded image.
      operationId: uploadProfilePhoto
      tags:
        - User
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                base64_image:
                  type: string
                  example: "data:image/png;base64,..."
      responses:
        "200":
          description: Profile photo uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwt:
                    type: string
                    example: "jwt.token.value"
                  message:
                    type: string
                    example: "Photo uploaded"
        "400":
          description: Bad request, no image uploaded
        "500":
          description: Internal server error
      security:
        - jwt: []
  /user:
    put:
      summary: Update User Profile
      description: Updates the user profile with provided details (firstname, lastname, email).
      operationId: updateUserProfile
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                  example: "John"
                lastname:
                  type: string
                  example: "Doe"
                email:
                  type: string
                  format: email
                  example: "johndoe@example.com"
                photo:
                  type: string
                  example: "base64encodedimage"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwt:
                    type: string
                    example: "jwt.token.value"
                  message:
                    type: string
                    example: "Profile updated"
        "400":
          description: Bad request, invalid profile data
        "404":
          description: User not found
        "500":
          description: Internal server error
      security:
        - jwt: []
  /user/password:
    patch:
      summary: Change User Password
      description: Changes the user's password, requires old password, new password, and confirmation.
      operationId: changeUserPassword
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                  example: "password123"
                newPassword:
                  type: string
                  example: "newpassword123"
                confirmPassword:
                  type: string
                  example: "newpassword123"
      responses:
        "200":
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password updated, login needed"
        "400":
          description: Bad request, missing passwords or mismatch
        "403":
          description: Unauthorized, old password incorrect
        "404":
          description: User not found
        "500":
          description: Internal server error
      security:
        - jwt: []
  /subscription:
    get:
      summary: Get User's Subscriptions
      description: Retrieves all subscriptions for the currently authenticated user.
      operationId: getUserSubscriptions
      tags:
        - Subscription
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
        "500":
          description: Internal server error
      security:
        - jwt: []
  /subscription/{id}:
    post:
      summary: Subscribe to a User
      description: Subscribes the currently authenticated user to another user by their ID.
      operationId: subscribeToUser
      tags:
        - Subscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 2
      responses:
        "201":
          description: Successfully subscribed to the user
        "500":
          description: Internal server error
        "400":
          description: Invalid target user ID
      security:
        - jwt: []

    delete:
      summary: Unsubscribe from a User
      description: Unsubscribes the currently authenticated user from another user by their ID.
      operationId: unsubscribeFromUser
      tags:
        - Subscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 2
      responses:
        "204":
          description: Successfully unsubscribed from the user
        "500":
          description: Internal server error
        "400":
          description: Invalid target user ID
      security:
        - jwt: []
...
  /room:
    get:
      summary: Get All Rooms for a User
      description: Retrieves all rooms owned by the authenticated user along with messages.
      operationId: getRooms
      tags:
        - Room
      responses:
        "200":
          description: List of all rooms for the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  rooms:
                    type: array
                    items:
                      $ref: "#/components/schemas/Room"
        "500":
          description: Internal server error
      security:
        - jwt: []

    post:
      summary: Create a New Room
      description: Creates a new room with the authenticated user as the owner.
      operationId: createRoom
      tags:
        - Room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "New Room"
      responses:
        "201":
          description: Room created successfully
        "409":
          description: Room already exists
        "500":
          description: Internal server error
      security:
        - jwt: []
  /room/{name}:
    get:
      summary: Get Room by Name
      description: Retrieves a room by its name along with its messages.
      operationId: getRoomByName
      tags:
        - Room
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
            example: "my-room"
      responses:
        "200":
          description: Room found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
        "500":
          description: Internal server error
      security:
        - jwt: []
  /room/add/{name}:
    post:
      summary: Add User to Room
      description: Adds a user to a room if the authenticated user is the room's owner.
      operationId: addUserToRoom
      tags:
        - Room
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
            example: "my-room"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                friendID:
                  type: integer
                  example: 2
      responses:
        "201":
          description: User added to room successfully
        "404":
          description: Room not found
        "403":
          description: Unauthorized to add user (not room owner)
        "500":
          description: Internal server error
      security:
        - jwt: []

    delete:
      summary: Delete a room
      description: Allows the authenticated user to delete a room by its ID.
      operationId: deleteRoom
      tags:
        - Room
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Room deleted successfully
        "404":
          description: Room not found
        "500":
          description: Internal server error
      security:
        - jwt: []
  /room/{id}:
    patch:
      summary: Update Room Name
      description: Updates the name of a room if the authenticated user is the owner of the room.
      operationId: patchRoomName
      tags:
        - Room
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Updated Room Name"
      responses:
        "200":
          description: Room name updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
        "403":
          description: Unauthorized to update room (not room owner)
        "500":
          description: Internal server error
      security:
        - jwt: []
    delete:
      summary: Delete a Room
      description: Deletes a room if the authenticated user is the owner of the room.
      operationId: deleteRoom
      tags:
        - Room
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Room deleted successfully
        "404":
          description: Room not found
        "403":
          description: Unauthorized to delete room (not room owner)
        "500":
          description: Internal server error
      security:
        - jwt: []
components:
  schemas:
...
    UserSummary:
      type: object
      properties:
        id:
          type: integer
          example: 100000
        firstname:
          type: string
          example: "John"
        lastname:
          type: string
          example: "Doe"
...
    Subscription:
      type: object
      xml:
        name: subscription
      properties:
        id:
          type: integer
          format: int64
          example: 100003
        ownerID:
          $ref: "#/components/schemas/UserSummary"
        targetID:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
          format: date-time
          example: "2024-10-22T10:15:30Z"
      required:
        - owner
        - target
        - createdAt
...
    Room:
      type: object
      xml:
        name: room
      properties:
        id:
          type: integer
          format: int64
          example: 100006
        name:
          type: string
          example: "General Discussion"
        ownerID:
          type: integer
          example: 100000
      required:
        - name
        - ownerID
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT